\name{simMAdata}
\docType{data}
\alias{simMAdata}
\title{Simulated Microarray data for testing purposes}
\description{
  This data set contains a mock microarray data set of 1000 genes and 60
  samples where the data has been generated from an underlying
  Independent Component Analysis model of 5 supergaussian modes.
}
\usage{data(simMAdata)}
\format{A list with three components. First, second and third components
  are the data matrix, mixing matrix and source matrix, respectively.}
\references{
  Hyvaerinen A., Karhunen J. and Oja E. (2001) \emph{Independent
    Component Analysis}. New York: Wiley.
}
\keyword{datasets}
